@JsonIgnore method to remove recursive callback wanna go in deep remove and check in postman

+++ src/main/java/com/aanchal/JobApplication/Jobs/Impl/jobServiceImpl.java
@@ -43,12 +43,11 @@ public class jobServiceImpl implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try{
+        if (jobRepository.existsById(id)) {
             jobRepository.deleteById(id);
             return true;
-        }catch (Exception ex){
-            return false;
         }
+        return false;
     }
 
     @Override
